Move schedule tests to the schedules app
